requirements.txt
numpy<2
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
faiss-cpu>=1.7.4
transformers>=4.31.0
streamlit>=1.24.0
yt-dlp>=2023.7.6